Increase MAX_INLINE from 20000 to 50000; update version to 2.2.5_gc
We have increased the number of lines that can be scanned from
the gckpp.kpp file from 20,000 to 50,000.  This is necessary as we have
been experimenting with splitting up rate law functions in order to
avoid having to use IF statements within functions (for efficiency).

Also updated the version number in gdata.h from 2.2.4_gc to 2.2.5_gc.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>