Remove calls to UPDATE_SUN and UPDATE_RCONST from rosenbrock*.f90
We have removed the calls to UPDATE_SUN() and UPDATE_RCONST() from
the int/rosenbrock.f90 and int/rosenbrock_adj.f90 files.

This will prevent these functions from being called within the
integration step, which is a performance hit.  In GEOS-Chem, we
call UPDATE_RCONST() outside of the integrator, in the DO_FLEXCHEM
routine.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>