Rename KPP/kpp_2.2.3_01 folder to KPP/kpp-code
The GC_updates branch is now at version 2.3.0_gc, which is several
updates ahead of the main branch (2.2.3_01).  This may lead to confusion
as to the current version for GEOS-Chem.

Therefore, we have renamed the KPP/kpp_2.2.3_01 folder to KPP/kpp-code,
which is independent of the version number.

The current version number for the GC_updates branch may be found
in KPP/kpp-code/gdata.h.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>