Restore Aout as an optional variable in Fun -- use direct writes
In order to have Aout be an optional variable in subroutine Fun,
we cannot use the normal FunctionBegin KPP function to write
the function interface and comments.  We now use direct writes
with the C-language fprintf statement to gckpp_Function.F90.

Also updated comments (mostly correcting the data) in code.f90
and gdata.h.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>