Rename master to main, update license and README.md
In keeping with Github best practices, we have renamed the master
branch to main.  We have also restored the GNU GPLv2 license
in the main branch.

We have also added kpp-code to the .gitignore file in the main
branch, as this folder name is only used in the GC_updates branch.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>