Replace "=>" token with "%" in F90-inlined code
We needed to implement a workaround in order to inline F90 code
with derived-type definitions.  The % character is interpreted
as a printf() format descriptor in C, so KPP cannot parse it
directly.

Added a string-replace function from StackOverflow to KPP routine
gen.c in order to replace the "=>" token with "%%%%".  After two
print statements, this will be written out to the F90 rate
module gckpp_Rates.F90 as %.

Therefore, any derived-type objects in the #INLINE section of
gckpp.kpp, such as:

   State_Het%xArea

must be written as

   State_Het=>xArea

in order for KPP to be able to properly parse it.

   State_Het%

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>