If ICNTRL(15)=1, then update rates in the Rosenbrock forward solver
This commit follows Github issue geoschem/KPP #5.  If the user sets
the ICNTRL(15) slot to 1, then the routine Update_Rconst() will be
called from within the Rosenbrock solver.  Normally this option will
be disabled.

We will modify the other solver files accordingly, but the Rosenbrock
solver is the only one used for GEOS-Chem, so we focus on this one
first.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>