Update CMakeLists to build the Flex/Bison generated code first
In the kpp-code/src/CMakeLists.txt, we now build library KppFlex,
which uses the outputs of the code generated by Bison and Flex
(i.e. y.tab.c, y.tab.h, lex.yy.c) first.

Then we build the rest of the KPP source code as the KppSource library,
which depends on KppFlex.  The executable depends on both libraries.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>