Copy source and header files to CMAKE_CURRENT_BINARY_DIR when building
The KPP build assumes that the output of the Parser (y.tab.c, y.tab.h)
and Lexer (lex.yy.c) will be placed in the source code folder.

The only way to build KPP properly without breaking the separation
between build output and source code is to copy all source code
and header files over to the CMAKE_CURRENT_BINARY_DIR, and then
tell the KppSource library to look for files there.

Also changed the permissions of source code and header files in
kpp-code/src.  These do not need to be executable, just read/write.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>